Build on stable Rust 1.98 with upstream serenity/songbird/poise
The three CycleFive forks in the workspace root were all dead weight:
the serenity mirror had no custom commits and was ~30 behind upstream,
the songbird mirror pointed at the identical upstream commit, and the
poise fork differed only by a stale `rev` pin. Since songbird 0.6 and
poise both resolve serenity from serenity-rs/serenity, keeping the
mirror forced two incompatible copies of serenity into the graph.

Also drops the [patch.crates-io.serenity-voice-model] entry, which
broke resolution outright now that the subcrate has left the serenity
repo, and moves crack-gpt from the async-openai fork (plus its backoff
patch) to upstream async-openai 0.41.

The only nightly feature in use was fmt_internals/formatting_options,
in a single test that Debug-formatted an embed, so rust-toolchain.toml
is now stable.

Migrates the API breakage that came with those bumps: the
ChannelId/GenericChannelId split, EventHandler's per-event methods
collapsing into a single dispatch, components v2, MessageUpdateEvent
becoming { message }, GuildChannel's flattened base, and poise
dropping FrameworkOptions::event_handler and usize command arguments.

+++ crack-core/src/commands/music/shuffle.rs
@@ -16,10 +16,10 @@ use rand::Rng;
 )]
 pub async fn movesong(
     ctx: Context<'_>,
-    #[description = "Index song is currently at"] at: usize,
-    #[description = "Index song will be moved to"] to: usize,
+    #[description = "Index song is currently at"] at: u32,
+    #[description = "Index song will be moved to"] to: u32,
 ) -> Result<(), Error> {
-    movesong_internal(ctx, at, to).await
+    movesong_internal(ctx, at as usize, to as usize).await
 }
 
 /// Move a song in the queue to a different position, internal function.
